requirements.txt
streamlit
pandas
openpyxl
python-docx
google-generativeai>=0.8.0
